Move timings implementation
