+++ src/main/java/co/aikar/timings/sponge/NullTimingHandler.java
@@ -22,30 +22,35 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.spongepowered.common.relocate.co.aikar.timings;
+package co.aikar.timings.sponge;
 
-import org.spongepowered.common.relocate.co.aikar.util.JSONUtil;
-import com.google.gson.JsonArray;
+import co.aikar.timings.Timing;
 
-class TimingHistoryEntry {
+public final class NullTimingHandler implements Timing {
 
-    final TimingData data;
-    final TimingData[] children;
-
-    TimingHistoryEntry(TimingHandler handler) {
-        this.data = handler.record.clone();
-        this.children = new TimingData[handler.children.size()];
-        int i = 0;
-        for (TimingData child : handler.children.values()) {
-            this.children[i++] = child.clone();
-        }
+    @Override
+    public Timing startTiming() {
+        return this;
     }
 
-    JsonArray export() {
-        JsonArray result = this.data.export();
-        if (this.children.length > 0) {
-            result.add(JSONUtil.mapArray(this.children, TimingData::export));
-        }
-        return result;
+    @Override
+    public void stopTiming() {
     }
+
+    @Override
+    public void startTimingIfSync() {
+    }
+
+    @Override
+    public void stopTimingIfSync() {
+    }
+
+    @Override
+    public void abort() {
+    }
+
+    @Override
+    public void close() {
+    }
+
 }
