+++ src/main/java/co/aikar/timings/sponge/SpongeTimings.java
@@ -22,7 +22,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.spongepowered.common.relocate.co.aikar.timings;
+package co.aikar.timings.sponge;
 
 import co.aikar.timings.Timing;
 import net.minecraft.world.entity.EntityType;
