+++ src/main/java/co/aikar/timings/sponge/TimingIdentifier.java
@@ -22,10 +22,10 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.spongepowered.common.relocate.co.aikar.timings;
+package co.aikar.timings.sponge;
 
 import co.aikar.timings.Timing;
-import org.spongepowered.common.relocate.co.aikar.timings.TimingIdentifier.TimingGroup;
+
 import org.spongepowered.common.relocate.co.aikar.util.LoadingMap;
 import org.spongepowered.common.relocate.co.aikar.util.MRUMapCache;
