+++ src/main/java/org/spongepowered/common/event/RegisteredListener.java
@@ -27,12 +27,11 @@ package org.spongepowered.common.event;
 import static com.google.common.base.Preconditions.checkNotNull;
 
 import co.aikar.timings.Timing;
-import com.google.common.collect.Lists;
-import com.google.common.collect.Maps;
+
 import org.spongepowered.api.event.Event;
 import org.spongepowered.api.event.EventListener;
 import org.spongepowered.api.event.Order;
-import org.spongepowered.common.relocate.co.aikar.timings.SpongeTimings;
+import co.aikar.timings.sponge.SpongeTimings;
 import org.spongepowered.plugin.PluginContainer;
 
 import java.util.ArrayList;
