+++ src/mixins/java/org/spongepowered/common/mixin/core/world/entity/EntityTypeMixin.java
@@ -28,7 +28,7 @@ import co.aikar.timings.Timing;
 import net.minecraft.world.entity.EntityType;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.common.bridge.world.entity.EntityTypeBridge;
-import org.spongepowered.common.relocate.co.aikar.timings.SpongeTimings;
+import co.aikar.timings.sponge.SpongeTimings;
 
 @Mixin(EntityType.class)
 public abstract class EntityTypeMixin implements EntityTypeBridge {
